Performance improvements.
No longer relying on Object.getPrototypeOf as it has large performance impact.

Operator overloads are now extracted from the object's constructor property.

Removed checks for number of arguments.

Removes construction of operator key inside of the operator callers. This was producing a large amount of strings at runtime.

Removed calls to Symbol.for as they were not adding value to maintainability but also had an impact on performance.